requirements.txt
# Requirements for developer machine
# Only select pre-compiled wheels for cryptography
# so we don't need Rust compiler installed on the 
# target device.
asn1crypto==1.5.1
awsiotsdk==1.13.0
bandit==1.7.5
boto3==1.26.113
cryptography==40.0.1
pylint==2.17.2
pytest==7.3.0
pytest-cov==4.0.0
pytest-mock==3.10.0
python-pkcs11==0.7.0
PyYAML==6.0
wheel
